Refactor API request handling
Use dedicated class/file for API request handling.
